requirements.txt
graphene-django
django-cors-headers 
django-filter 
bcrypt